create mistakes from mentor's

+++ controllers/contactsControllers.js
@@ -1,5 +1,8 @@
 import HttpError from '../helpers/HttpError.js';
-import { Contact } from '../schemas/contactsSchemas.js';
+import {
+  createContactSchema,
+  updateContactSchema,
+} from '../schemas/contactsSchemas.js';
 
 export const getAllContacts = async (_, res, next) => {
   try {
@@ -13,9 +16,13 @@ export const getAllContacts = async (_, res, next) => {
 export const getOneContact = async (req, res, next) => {
   try {
     const { id } = req.params;
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      throw new HttpError(400, 'Invalid ID');
+    }
+
     const result = await Contact.findById(id);
     if (!result) {
-      throw HttpError(404, 'Not found');
+      throw new HttpError(404, 'Not found');
     }
     res.status(200).json(result);
   } catch (error) {
@@ -26,16 +33,20 @@ export const getOneContact = async (req, res, next) => {
 export const deleteContact = async (req, res, next) => {
   try {
     const { id } = req.params;
+
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      throw new HttpError(400, 'Invalid ID');
+    }
+
     const result = await Contact.findByIdAndDelete(id);
     if (!result) {
-      throw HttpError(404, 'Not found');
+      throw new HttpError(404, 'Not found');
     }
     res.status(200).json(result);
   } catch (error) {
     next(error);
   }
 };
-
 export const createContact = async (req, res, next) => {
   try {
     const contact = {
@@ -61,19 +72,23 @@ export const updateContact = async (req, res, next) => {
   try {
     const { id } = req.params;
     const updatedData = req.body;
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      throw new HttpError(400, 'Invalid ID');
+    }
+
     if (Object.keys(updatedData).length === 0) {
-      throw HttpError(400, 'Body must have at least one field');
+      throw new HttpError(400, 'Body must have at least one field');
     }
     const { error } = updateContactSchema.validate(updatedData);
     if (error) {
-      throw HttpError(400, error.message);
+      throw new HttpError(400, error.message);
     }
     const result = await Contact.findByIdAndUpdate(id, updatedData, {
       new: true,
       runValidators: true,
     });
     if (!result) {
-      throw HttpError(404, 'Not found');
+      throw new HttpError(404, 'Not found');
     }
     res.status(200).json(result);
   } catch (error) {
@@ -85,8 +100,13 @@ export const updateStatusContact = async (req, res, next) => {
   try {
     const { id } = req.params;
     const { favorite } = req.body;
+
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      throw new HttpError(400, 'Invalid ID');
+    }
+
     if (favorite === undefined) {
-      throw HttpError(400, 'Missing field favorite');
+      throw new HttpError(400, 'Missing field favorite');
     }
     const result = await Contact.findByIdAndUpdate(
       id,
@@ -94,7 +114,7 @@ export const updateStatusContact = async (req, res, next) => {
       { new: true, runValidators: true }
     );
     if (!result) {
-      throw HttpError(404, 'Not found');
+      throw new HttpError(404, 'Not found');
     }
     res.status(200).json(result);
   } catch (error) {
